+++ SallesWebMvc/SallesWebMvc/Models/Seller.cs
@@ -6,7 +6,7 @@
         public string Name { get; set; }
         public string Email { get; set; }
         public DateTime BirthDate { get; set; }
-        public double BaseSalary { get; set; }
+        public decimal BaseSalary { get; set; }
         public Department? Department { get; set; }
         public int DepartmentId { get; set; }
         public ICollection<SellesRecord> Salles { get; set; } = new List<SellesRecord>();
@@ -15,7 +15,7 @@
         {
         }
 
-        public Seller(int id, string name, string email, DateTime birthDate, double baseSalary, Department department)
+        public Seller(int id, string name, string email, DateTime birthDate, decimal baseSalary, Department department)
         {
             Id = id;
             Name = name;
@@ -35,7 +35,7 @@
             Salles.Remove(sallesRecord);
         }
 
-        public double TotalSalles(DateTime initial, DateTime final)
+        public decimal TotalSalles(DateTime initial, DateTime final)
         {
             return Salles.Where(sallesRecord => sallesRecord.Date >= initial && sallesRecord.Date <= final).Sum(sallesRecord => sallesRecord.Amount);
         }
